Continuing to abstract materials
Changing draw code to use abstraction, does not work like we thought it
would. Committing to work from campus tomorrow. For some fucking reason,
the Wall Material uses the player pixel shader. The program will crash
the video drivers is pixelShader->SetShader() is called in WallMaterial.
Literally not a clue why yet. Currently it is commented out and the game
runs, just with no depth lines.