Virtualized Shader Preparation, Condensed Draw Code
Virtualized the PrepareToDraw method in Material, making the method
overrideable and able to be called via the base class. This allowed me
to significantly condense the code for drawing in order of Material.
This still doesn't solve the WallMaterial pixel shader issue though, I
have no idea what is up with that. I also changed the clear color to a
dark grey. The Walls will render using the player pixel shader for now,
because the wall pixel shader never gets set. This is a problem because
it's signature for data doesn't match the vertex shader, but at least
the game runs fine. Will fix in class tomorrow.